fix(db-schema): personal and shared expenses
In order to be able to distinguish between personal and shared expenses, two
new tables are created. They share an underlying reference to an expense
record, which hold very little information, but is common to both types of
expenses.

This will allow users to track their own expenses and also share expenses in
groups with other members. One could think that creating a group for yourself
would be a solution, but it feels hacky, so we are going with this approach,
let's see if it works out.