fix(theme): something broke the settlement pop-up when switching to 'primary'
I don't know what broke it, but I reverted the change and did it again. I know
I did not run the text substitutions in the best way possible, so I just did it
again doing what I think is the best way, and this time it seems nothing broke.