feat: sync step when checking users
In preparation for data migrations (moving from platform to platform, or
importing prod data into dev), we uniquely identify users by email when
comparing with clerk data. And if the reference in clerk does not match our
internal reference, then we issue an update of the data in clerk.

That will, in turn, trigger a 'user.updated' webhook, which will then sync
things backs into our database.